config: move ConfigResultExt from settings module
As the name suggests, this helper should be in the config module.
