+++ cli/src/commands/git/fetch.rs
@@ -14,8 +14,8 @@
 
 use clap_complete::ArgValueCandidates;
 use itertools::Itertools;
+use jj_lib::config::ConfigResultExt as _;
 use jj_lib::repo::Repo;
-use jj_lib::settings::ConfigResultExt as _;
 use jj_lib::settings::UserSettings;
 use jj_lib::str_util::StringPattern;
