+++ lib/src/ssh_signing.rs
@@ -27,7 +27,7 @@ use either::Either;
 use thiserror::Error;
 
 use crate::config::ConfigError;
-use crate::settings::ConfigResultExt as _;
+use crate::config::ConfigResultExt as _;
 use crate::settings::UserSettings;
 use crate::signing::SigStatus;
 use crate::signing::SignError;
